requirements.txt
dateparser==1.0.0
psutil==5.8.0
pygit2==1.5.0
requests==2.22.0
pyfiglet==0.8.post1
jishaku==1.20.0.220
parsedatetime==2.6
dpymenus==1.3.1
aiohttp==3.7.4.post0
disputils==0.1.3
ksoftapi==0.4.0
pygicord==0.4.1
asyncpg==0.22.0
discord==1.0.1
discord.py==1.7.2
python_dateutil==2.8.1
dblpy==0.4.0
mystbin.py
buttons==0.1.9